Remove Blobs from news endpoints
